tests/wpt/web-platform-tests/navigation-api/state/same-document-away-and-back-location-api.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(async () => {
    let entry0 = navigation.currentEntry;

    let navState = { statevar: "state" };
    await navigation.navigate("#1", { state: navState });
    let entry1 = navigation.currentEntry;

    location.href = "#2";
    let entry2 = navigation.currentEntry;

    assert_equals(navigation.entries().length, 3);
    assert_equals(entry0, navigation.entries()[0]);
    assert_equals(entry1, navigation.entries()[1]);
    assert_equals(entry2, navigation.entries()[2]);

    assert_equals(entry0.getState(), undefined);

    let state1 = entry1.getState();
    assert_not_equals(state1, undefined);
    assert_not_equals(state1, navState);
    assert_equals(state1.statevar, "state");

    let state2 = entry2.getState();
    assert_not_equals(state2, undefined);
    assert_not_equals(state2, navState);
    assert_equals(state2.statevar, "state");

    history.back();
    window.onpopstate = t.step_func_done(() => {
      assert_equals(navigation.entries().length, 3);
      let back_entry = navigation.currentEntry;
      assert_equals(back_entry, navigation.entries()[1]);
      let back_state = back_entry.getState();
      assert_not_equals(back_state, state1);
      assert_not_equals(back_state, state2);
      assert_equals(back_state.statevar, "state");
    });
  }, 0);
}, "entry.getState() behavior after navigating away using a fragment change, then back");
</script>
